Extract clean filter function logic
